pyproject.toml
[build-system]
requires = ["flit_core >=2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "rice"
authors = [
	{ name = "Hayley Hughes", email  = "hayley@foxes.systems" },
]
classifiers = ["License :: OSI Approved :: Apache Software License"]
readme = "README.md"
requires-python=">=3.6"
dynamic = ["version", "description"]
dependencies = [
	"jinja2",
	"docopts",
]

[project.urls]
Source = "https://github.com/hugglesfox/rice"

[project.scripts]
rice = "rice.__main__"

